+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # 1.4.1
+* FEATURE: Display number of element for each category
 * MINOR: Improve UI (theme, fontawesome icons)
 
 #1.4